requirements.txt
fastapi
uvicorn
firebase_admin
pydantic_settings
langchain-community
unstructured
markdown
openai
langchain-openai
faiss-cpu
python-dotenv
pandas
langgraph
langchain_anthropic
scikit-learn
langchain_experimental
langchain_chroma
firebase-admin
